feat(c-api) Move all types into their own modules.
