+++ lib/c-api/src/wasm_c_api/macros.rs
@@ -155,3 +155,21 @@ macro_rules! wasm_declare_own {
         }
     };
 }
+
+#[macro_export]
+macro_rules! c_try {
+    ($expr:expr) => {{
+        let res: Result<_, _> = $expr;
+        match res {
+            Ok(val) => val,
+            Err(err) => {
+                crate::error::update_last_error(err);
+                return None;
+            }
+        }
+    }};
+    ($expr:expr, $e:expr) => {{
+        let opt: Option<_> = $expr;
+        c_try!(opt.ok_or_else(|| $e))
+    }};
+}
