+++ lib/c-api/src/wasm_c_api/types/byte.rs
@@ -1,0 +1,4 @@
+#[allow(non_camel_case_types)]
+pub type wasm_byte_t = u8;
+
+wasm_declare_vec!(byte);
