+++ lib/c-api/src/wasm_c_api/types/frame.rs
@@ -1,0 +1,5 @@
+#[derive(Debug)]
+#[repr(C)]
+pub struct wasm_frame_t {}
+
+wasm_declare_vec!(frame);
