+++ lib/c-api/src/wasm_c_api/types/mod.rs
@@ -1,0 +1,27 @@
+mod byte;
+mod export;
+mod extern_;
+mod frame;
+mod function;
+mod global;
+mod import;
+mod memory;
+mod mutability;
+mod name;
+mod reference;
+mod table;
+mod value;
+
+pub use byte::*;
+pub use export::*;
+pub use extern_::*;
+pub use frame::*;
+pub use function::*;
+pub use global::*;
+pub use import::*;
+pub use memory::*;
+pub use mutability::*;
+pub use name::*;
+pub use reference::*;
+pub use table::*;
+pub use value::*;
