+++ lib/c-api/src/wasm_c_api/types/name.rs
@@ -1,0 +1,4 @@
+use super::wasm_byte_vec_t;
+
+#[allow(non_camel_case_types)]
+pub type wasm_name_t = wasm_byte_vec_t;
