+++ lib/c-api/src/wasm_c_api/types/reference.rs
@@ -1,0 +1,3 @@
+// opaque type over `ExternRef`?
+#[allow(non_camel_case_types)]
+pub struct wasm_ref_t;
